Add MIPS r6 support
MIPS r6 changes the multiply instructions
The previous version use non-GPR for reuslt, while r6 use GPR.

The instructions are also replaced by new one:
  multu -> mulu/muhu
  dmultu -> dmulu/dmuhu

We rewrite the code for MIPS r6, to make it in a single loop.
Then the length of code is quite shorter now.
With some test, the performance is almost same with the privious.